Update of lab 1 (job market) and lab 2 exercises
Lab 2 exercises - I realised as well that in lab 1 the job market jupyter notebook file that I created was saved in another directory (the one that I used to save my jupyter notebook files). I know that it's late but I submit anyway as a proof that I had done the exercise.